fix payload null + add data mitra + rating for job by customer + fix validate etc

+++ src/jobs/jobs.controller.js
@@ -1,5 +1,6 @@
 const Boom = require('@hapi/boom');
 const jobsService = require('../jobs/jobs.service');
+const mitrasService = require('../mitra/mitras.service');
 const customersModule = require('../customer/customers.module');
 const mitrasModule = require('../mitra/mitras.module');
 const transactionsModule = require('../transaction/transactions.module');
@@ -8,6 +9,14 @@ const { getCurrentTime, getTime } = require('../utils/time');
 const { uploadFileToCloudStorage } = require('../server/storage');
 
 const createJobHandler = async (request, h) => {
+    if (!request.payload || Object.keys(request.payload).length === 0) {
+        const boomError = Boom.badRequest('Request payload cannot be empty.');
+        return h.response({
+            status: boomError.output.statusCode,
+            message: boomError.message,
+            error: true,
+        }).code(boomError.output.statusCode);
+    }
     try {
         const allowedParams = ['title', 'deadline', 'location', 'cost', 'description', 'image'];
         const payloadKeys = Object.keys(request.payload);
@@ -42,7 +51,7 @@ const createJobHandler = async (request, h) => {
             }).code(boomError.output.statusCode);
         }
 
-        const currentDate = getTime(getCurrentTime()); 
+        const currentDate = getTime(getCurrentTime());
         const providedDeadline = getTime(deadline);
         currentDate.setHours(0, 0, 0, 0);
         providedDeadline.setHours(0, 0, 0, 0);
@@ -182,6 +191,43 @@ const deleteJobHandler = async (request, h) => {
 
         const { customer_id } = customer;
 
+        const job = await jobsService.getJobById(job_id);
+        if (!job) {
+            const boomError = Boom.notFound('Job not found, Please check your job id.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (job.status == 'Canceled') {
+            const boomError = Boom.badRequest('Job is has been canceled, Please check other jobs.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (job.status == 'Completed') {
+            const boomError = Boom.badRequest('Job is has been completed, Please check other jobs.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (job.status != 'Pending') {
+            const boomError = Boom.badRequest('Job cannot be deleted, Please try again.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
         await jobsService.deleteJobById(job_id, customer_id);
 
         return h.response({
@@ -201,6 +247,14 @@ const deleteJobHandler = async (request, h) => {
 };
 
 const updateJobHandler = async (request, h) => {
+    if (!request.payload || Object.keys(request.payload).length === 0) {
+        const boomError = Boom.badRequest('Request payload cannot be empty.');
+        return h.response({
+            status: boomError.output.statusCode,
+            message: boomError.message,
+            error: true,
+        }).code(boomError.output.statusCode);
+    }
     try {
         const { job_id } = request.params;
         const { userId } = request.auth.credentials;
@@ -248,8 +302,8 @@ const updateJobHandler = async (request, h) => {
             }).code(boomError.output.statusCode);
         }
 
-        if (job.status == 'In Progress') {
-            const boomError = Boom.forbidden('Job cannot be edited because it is already in progress.');
+        if (job.status == 'Canceled') {
+            const boomError = Boom.badRequest('Job is has been canceled, Please check other jobs.');
             return h.response({
                 status: boomError.output.statusCode,
                 message: boomError.message,
@@ -258,7 +312,16 @@ const updateJobHandler = async (request, h) => {
         }
 
         if (job.status == 'Completed') {
-            const boomError = Boom.forbidden('Job cannot be edited because it is already completed.');
+            const boomError = Boom.badRequest('Job is has been completed, Please check other jobs.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (job.status != 'Pending') {
+            const boomError = Boom.forbidden('Job is has been in progress, Please check other jobs.');
             return h.response({
                 status: boomError.output.statusCode,
                 message: boomError.message,
@@ -374,6 +437,15 @@ const assignJobHandler = async (request, h) => {
             }).code(boomError.output.statusCode);
         }
 
+        if (job.status == 'Completed') {
+            const boomError = Boom.badRequest('Job is has been completed, Please check other jobs.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
         if (job.status != 'Pending') {
             const boomError = Boom.badRequest('Job is has been taken, Please check other jobs.');
             return h.response({
@@ -481,6 +553,24 @@ const completeJobHandler = async (request, h) => {
             }).code(boomError.output.statusCode);
         }
 
+        if (job.status == 'Completed') {
+            const boomError = Boom.badRequest('Job cannot marked completed, Your job already completed.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (job.status == 'Canceled') {
+            const boomError = Boom.badRequest('Job cannot marked completed, Your job canceled.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
         if (job.status != 'In Progress') {
             const boomError = Boom.badRequest('Job cannot marked completed, Your job not in progress.');
             return h.response({
@@ -491,6 +581,10 @@ const completeJobHandler = async (request, h) => {
         }
 
         const updatedJob = await jobsService.completeJob(job_id);
+
+        if (job.mitra_id) {
+            await mitrasService.updateTransactionDone(job.mitra_id);
+        }
 
         return h.response({
             status: 200,
@@ -597,6 +691,111 @@ const getJobDetailHandler = async (request, h) => {
     }
 };
 
+const addRatingHandler = async (request, h) => {
+    if (!request.payload || Object.keys(request.payload).length === 0) {
+        const boomError = Boom.badRequest('Request payload cannot be empty.');
+        return h.response({
+            status: boomError.output.statusCode,
+            message: boomError.message,
+            error: true,
+        }).code(boomError.output.statusCode);
+    }
+    try {
+        const { job_id } = request.params;
+        const { rating } = request.payload;
+        const { userId } = request.auth.credentials;
+
+        const allowedParams = ['rating'];
+        const payloadKeys = Object.keys(request.payload);
+        const invalidParams = payloadKeys.filter((key) => !allowedParams.includes(key));
+        if (invalidParams.length > 0) {
+            const boomError = Boom.badRequest(`/ ${invalidParams.join(', ')} / not allowed.`);
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        const customer = await customersModule.findCustomerByUserId(userId);
+        if (!customer) {
+            const boomError = Boom.notFound('Customer profile not found, Please try again.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (!rating || isNaN(rating) || rating < 1 || rating > 5) {
+            const boomError = Boom.badRequest('Rating must be a number between 1 and 5.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        const job = await jobsService.getJobById(job_id);
+        if (!job) {
+            const boomError = Boom.notFound('Job not found, Please check your job id.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (job.rating) {
+            const boomError = Boom.badRequest('Rating has already been added to this job.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        const { customer_id } = customer;
+
+        if (job.customer_id != customer_id) {
+            const boomError = Boom.forbidden('Access denied, You do not have permission to rating this job.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        if (job.status != 'Completed') {
+            const boomError = Boom.badRequest('Rating can only be given for completed jobs, Please try again.');
+            return h.response({
+                status: boomError.output.statusCode,
+                message: boomError.message,
+                error: true,
+            }).code(boomError.output.statusCode);
+        }
+
+        const updatedJob = await jobsService.addRatingToJob(job_id, rating);
+
+        if (job.mitra_id) {
+            await mitrasService.updateMitraRating(job.mitra_id);
+        }
+
+        return h.response({
+            status: 200,
+            message: 'Rating added successfully',
+            error: false,
+        }).code(200);
+    } catch (error) {
+        const boomError = Boom.badRequest(error.message);
+        return h.response({
+            status: boomError.output.statusCode,
+            message: boomError.message,
+            error: true,
+        }).code(boomError.output.statusCode);
+    }
+};
+
 module.exports = {
     createJobHandler,
     getJobsHandler,
@@ -608,5 +807,6 @@ module.exports = {
     cancelOverdueJobsHandler,
     getPendingJobsHandler,
     getAllJobsHandler,
-    getJobDetailHandler
+    getJobDetailHandler,
+    addRatingHandler
 };
